cpukit/.../scheduler*: Address -Wsign-compare warnings
This warning occurs when comparing a signed variable to an unsigned one.
This is frequently an int or ssize_t variable compared to a uint32_t or
size_t. Sometimes the size_t is from a sizeof() use.

This addresses sign comparison warnings primarily in the CBS.

+++ cpukit/rtems/src/schedulermapfromposix.c
@@ -62,7 +62,7 @@ rtems_status_code rtems_scheduler_map_priority_from_posix(
     return RTEMS_INVALID_PRIORITY;
   }
 
-  if ( posix_priority >= scheduler->maximum_priority ) {
+  if ( (uint32_t) posix_priority >= scheduler->maximum_priority ) {
     return RTEMS_INVALID_PRIORITY;
   }
